refactor: reorganizacion de componentes
