+++ dev-dist/sw.js
@@ -86,7 +86,7 @@ define(['./workbox-84006cf1'], (function (workbox) { 'use strict';
     "revision": "3ca0b8505b4bec776b69afdba2768812"
   }, {
     "url": "index.html",
-    "revision": "0.o3s01l04dk"
+    "revision": "0.j7ebfsnt498"
   }], {});
   workbox.cleanupOutdatedCaches();
   workbox.registerRoute(new workbox.NavigationRoute(workbox.createHandlerBoundToURL("index.html"), {
