+++ src/components/common/errors/SWErrorHandler.jsx
