+++ src/components/common/index.js
@@ -1,13 +1,28 @@
-export { default as Button } from './Button';
-export { default as Card } from './Card';
-export { default as ErrorMessage } from './ErrorMessage';
-export { default as LoadingSpinner } from './LoadingSpinner';
-export { default as ProgressBar } from './ProgressBar';
-export { default as Dialog } from './Dialog';
-export { default as ServiceWorkerUpdater } from './ServiceWorkerUpdater';
-export { default as SWErrorHandler } from './SWErrorHandler';
-export { default as Toast } from './Toast';
-export { default as ToastContainer } from './ToastContainer';
-export { default as CodeBlock } from './CodeBlock';
-export { default as DOMRecoveryMonitor } from './DOMRecoveryMonitor';
-export { default as GlobalErrorBoundary } from './GlobalErrorBoundry';
+// Botones
+export { default as Button } from './buttons/Button';
+export { default as FavoriteButton } from './buttons/FavoriteButton';
+
+// Cards y contenedores visuales
+export { default as Card } from './cards/Card';
+
+// Diálogos y modales
+export { default as Dialog } from './dialogs/Dialog';
+export { default as Modal } from './dialogs/Modal';
+
+// Feedback y notificaciones
+export { default as ErrorMessage } from './feedback/ErrorMessage';
+export { default as LoadingSpinner } from './feedback/LoadingSpinner';
+export { default as ProgressBar } from './feedback/ProgressBar';
+export { default as Toast } from './feedback/Toast';
+export { default as ToastContainer } from './feedback/ToastContainer';
+
+// Manejo de errores y recuperación
+export { default as GlobalErrorBoundary } from './errors/GlobalErrorBoundary';
+export { default as DOMRecoveryMonitor } from './errors/DOMRecoveryMonitor';
+export { default as SWErrorHandler } from './errors/SWErrorHandler';
+
+// Service Worker
+export { default as ServiceWorkerUpdater } from './sw/ServiceWorkerUpdater';
+
+// Componentes de código y recursos
+export { default as CodeBlock } from './code/CodeBlock';
