+++ src/components/quiz/media/ImageResource.jsx
