+++ src/components/quiz/pending/PendingQuizzes.jsx
@@ -8,8 +8,7 @@
  */
 import { useState, useEffect, useCallback } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { Button } from '@components/common';
-import Modal from '@components/common/Modal';
+import { Button, Modal } from '@components/common';
 import { formatTimestamp } from '@utils/quizUtils';
 
 export default function PendingQuizzes() {
