+++ src/components/quiz/questions/QuestionReview.jsx
@@ -20,8 +20,8 @@
  */
 import { useState, useEffect } from 'react';
 import { useTheme, useDeviceType } from '@hooks';
-import ImageResource from './ImageResource';
-import FavoriteButton from './FavoriteButton';
+import ImageResource from '../media/ImageResource';
+import FavoriteButton from '../favorites/FavoriteButton';
 import { addFavorite, removeFavorite } from '@services/favoritesService';
 import { showToast } from '@utils/toastUtils';
 import { getQuestionReviewStyles } from '@styles/safeStyles';
